fix: environment-dependent build errors in lvgl font path
Three issues surfaced on the Seeed Wio Terminal (SAMD51, Arduino)
build, all triggered by the lvgl font feature sync:

1. Arduino.h defines abs() as a preprocessor macro, which collides
   with std::abs used inside lgfx_fonts.cpp. Add abs to the existing
   min/max #undef block at the top of the file so std::abs resolves
   cleanly on every toolchain.

2. The switch on gd.format in LVGLfont::drawChar was guarded with
   #ifdef LV_FONT_GLYPH_FORMAT_A1 etc, but these identifiers are enum
   values defined in lv_font/font.h, not preprocessor macros - so the
   guards always evaluated false, every case body became unreachable,
   and glyph_bpp was never updated. Drop the broken #ifdef guards and
   use the enum labels directly.

3. fill_rect_radial_gradient / fill_rect_linear_gradient in LGFXBase
   compared int loop counters against uint32_t dimensions, producing
   -Wsign-compare warnings. Cast the upper bounds to int so the loop
   variables keep their signed type (leaving downstream x/y pushImage
   arithmetic unchanged).

+++ src/lgfx/v1/LGFXBase.cpp
@@ -1093,9 +1093,9 @@ namespace lgfx
       auto scanline = (rgb888_t*)alloca(w * sizeof(rgb888_t));
 
       startWrite();
-      for( int _y=0;_y<h;_y++ ) {
+      for( int _y=0;_y<(int)h;_y++ ) {
         // only half of the scan line needs to be calculated, the other half is mirrored
-        for( int _x=0;_x<=w/2;_x++ ) {
+        for( int _x=0;_x<=(int)w/2;_x++ ) {
           auto distance       = pixelDistance( fmidx, fmidy, _x*vratio, _y*hratio );
           scanline[_x]        = map_gradient( distance, 0, hyp0, gradient );
           scanline[(w-1)-_x]  = scanline[_x];
@@ -1118,11 +1118,11 @@ namespace lgfx
     bool is_vertical = style==VLINEAR;
     const uint32_t gradient_len = is_vertical ? h : w;
     auto scanline = (rgb888_t*)alloca(gradient_len * sizeof(rgb888_t));
-    for(int i=0;i<gradient_len;i++) { // memoize one gradient scanline
+    for(int i=0;i<(int)gradient_len;i++) { // memoize one gradient scanline
       scanline[i] = map_gradient( i, 0, gradient_len, gradient );
     }
     startWrite();
-    for( int ys=0;ys<h;ys++ ) {
+    for( int ys=0;ys<(int)h;ys++ ) {
       if( is_vertical ) { // scanline is used as an colors index
         setColor(color888(scanline[ys].r, scanline[ys].g, scanline[ys].b));
         drawFastHLine( x, y+ys, w );
